Add endpoint, GRPC transport and main to polling
The GRPC transport and endpoint for Go Kit haven been added with this
commit.

Furthermore, the package name has been changed in order to create a Go
command rather than a Go package.

A makefile has been added with the possibility to build the polling
service with the vgo prototype.

The fridgelethics-messages submodule has been removed, as the required
Go code is now directly imported from the repository.